requirements.txt
anthropic>=0.39.0
playwright>=1.48.0
feedparser>=6.0.11
httpx>=0.27.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pyyaml>=6.0.2
tavily-python>=0.5.0  # Web検索API（オプション）
